server/requirements.txt
boto3==1.40.35
botocore==1.40.35
jmespath==1.0.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
